--- components/DashboardList.tsx
@@ -1,70 +1,0 @@
-import { getDashboard } from '@/pages/api/getDashboard';
-import { Dashboard, DashboardResponse } from '@/types/myDashboardTypes';
-import Image from 'next/image';
-import { useEffect, useState } from 'react';
-
-
-
-import MyDashboardItem from './MyDashboardItem';
-import PaginationBar from './PaginationBar';
-
-
-const DashboardList: React.FC = () => {
-  const [page, setPage] = useState(1);
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [size, setSize] = useState(5);
-  const [allDashboardList, setAllDashboardList] = useState<Dashboard[]>([]);
-  const [totalPage, setTotalPage] = useState(1);
-  useEffect(() => {
-    const getDashboardData = async () => {
-      try {
-        const getDashboardRes: DashboardResponse = await getDashboard({
-          page,
-          size,
-        });
-        setAllDashboardList(getDashboardRes.dashboards);
-        setTotalPage(Math.ceil(getDashboardRes.totalCount / size));
-      } catch (err) {
-        console.log(err);
-      }
-    };
-    getDashboardData();
-  }, [page, size]);
-
-  const onPageChange = (pageNumber: number) => {
-    setPage(pageNumber);
-  };
-
-  return (
-    <>
-      <ul className="container mx-auto grid h-fit w-fit grid-cols-1 grid-rows-6 gap-y-8">
-        <div className="flex h-58 w-260 flex-row items-center justify-between rounded-lg border border-solid border-gray-30 px-56 py-23">
-          <p className="text-sm font-semibold">새로운 대시보드</p>
-          <div className="relative h-22 w-22">
-            <Image
-              fill
-              src="/images/icon/ic-color-add.svg"
-              alt="새로운 대시보드 만들기"
-            />
-          </div>
-        </div>
-        {allDashboardList &&
-          allDashboardList.map((myDashboard) => {
-            return (
-              <li className="w-fit" key={myDashboard.id}>
-                <MyDashboardItem myDashboard={myDashboard} />
-              </li>
-            );
-          })}
-        {allDashboardList && (
-          <PaginationBar
-            totalPage={totalPage}
-            activePage={page}
-            onPageChange={onPageChange}
-          />
-        )}
-      </ul>
-    </>
-  );
-};
-export default DashboardList;
