+++ components/SideBar.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const SideBar=()=>{
+    return(
+        <aside>사이드바</aside>
+    )
+}
+export default SideBar;
